knowledge-graphs/pyproject.toml
[project]
name = "knowledge-graphs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.27.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.8",
    "langchain-neo4j>=0.4.0",
    "langchain-qdrant>=0.2.0",
    "mem0ai>=0.1.114",
    "python-dotenv>=1.1.1",
]
